+++ bundles/subclipse.ui/src/org/tigris/subversion/subclipse/ui/compare/SVNLocalCompareSummaryInput.java
@@ -63,7 +63,7 @@ public class SVNLocalCompareSummaryInput extends SVNAbstractCompareEditorInput
   private void initLabels() {
     CompareConfiguration cc = getCompareConfiguration();
     cc.setLeftEditable(!readOnly);
-    cc.setRightEditable(false);
+    cc.setRightEditable(!readOnly);
     String title;
     String leftLabel;
     String rightLabel;